+++ packages/core/src/index.ts
@@ -2,6 +2,7 @@ export * from './roxie';
 export * from './ajax';
 export * from './observable';
 export * from './data';
+export * from './types';
 
 export const logInfo = (msg: string) => console.log(`[INFO] ${msg}`);
 export const logWarn = (msg: string) => console.warn(`[WARN] ${msg}`);