+++ packages/ui/src/roxie-ui.ts
@@ -1,8 +1,8 @@
-import 'jquery';
+import 'jquery/dist/jquery.slim';
 import 'popper.js';
 import 'bootstrap';
 
-import $ from 'jquery';
+import $ from 'jquery/dist/jquery.slim';
 
 $(function() {
   $('[data-toggle="tooltip"]').tooltip();